+++ R/review_pull_requests.R
@@ -1,6 +1,6 @@
 #' @title Review R-multiverse contribution pull requests.
 #' @export
-#' @family package reviews
+#' @family Automated package reviews
 #' @description Review pull requests which add packages to `packages.json`.
 #' @return `NULL` (invisibly).
 #' @inheritParams review_pull_request